+++ test/features/dictionary_test.exs
@@ -8,6 +8,7 @@ defmodule TdDd.DictionaryTest do
 
   alias Jason, as: JSON
   alias TdDd.DataStructures.PathCache
+  alias TdDd.Lineage.GraphData
   alias TdDd.Loader.LoaderWorker
   alias TdDd.Permissions.MockPermissionResolver
   alias TdDd.Search.MockIndexWorker
@@ -43,6 +44,7 @@ defmodule TdDd.DictionaryTest do
     start_supervised(MockIndexWorker)
     start_supervised(LoaderWorker)
     start_supervised(PathCache)
+    start_supervised(GraphData)
     :ok
   end